requirements.txt
fastapi~=0.111.0
requests~=2.32.3
pydantic~=2.7.4
PyYAML~=6.0.1
APScheduler~=3.10.4
starlette~=0.37.2
cachetools~=5.3.3
aiohttp~=3.9.5
passlib~=1.7.4
PyJWT~=2.8.0
bcrypt~=4.1.3